requirements.txt
# RAG and Langchain Related
langchain
langchain-community
sentence-transformers
langchain-google-genai
faiss-cpu
openai
deeplake

# Document Parsing
PyPDF2

# Utils
python-dotenv

# UI
streamlit
streamlit-extras

# Google Drive Access
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
